Pequenas alterações no código do CUBE e nas MTS em Python

+++ CubeIDE/MTS_f7/Core/Inc/calib.h
@@ -8,11 +8,13 @@
 #ifndef INC_CALIB_H_
 #define INC_CALIB_H_
 
+#include "stm32f7xx_hal.h"
+
 #define N 9
 #define tam 1111
 
 void ETS(float *mx, float *my, float *mz, float *p);
-int NLLS(float *mx, float *my, float *mz, float *p);
+uint8_t NLLS(float *mx, float *my, float *mz, float *p);
 
 void inverter_matriz(float matriz[N][N], float inversa[N][N]);
